made changes suggested
