+++ website/docs/r/macie2_invitation_accepter.html.markdown
@@ -53,8 +53,8 @@ In addition to all arguments above, the following attributes are exported:
 
 ## Import
 
-`aws_macie2_invitation_accepter` can be imported using the id, e.g.
+`aws_macie2_invitation_accepter` can be imported using the admin account ID, e.g.
 
 ```
-$ terraform import aws_macie2_invitation_accepter.example abcd1
+$ terraform import aws_macie2_invitation_accepter.example 123456789012
 ```
